requirements.txt
openai==0.27.0
pandas==1.2.0
numpy==1.19.0
scikit-learn==0.24.0
torch==1.8.0
pytest==6.0.0
pyyaml==5.3.0
